tools/simulator/pyproject.toml
[project]
name = "waffleweather-simulator"
version = "0.1.0"
description = "Weather data simulator for WaffleWeather development"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "paho-mqtt>=2.0",
    "httpx>=0.27",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
]

[project.scripts]
simulator = "simulator:cli"
